+++ Blueprint-Core/BlueprintCore/Blueprints/Components/ContextRankConfigs.cs
@@ -1,3 +1,4 @@
+using BlueprintCore.Utils;
 using Kingmaker.Blueprints;
 using Kingmaker.EntitySystem.Stats;
 using Kingmaker.Enums;