+++ BlueprintCoreGen/CodeGen/FieldProcessor.cs
@@ -1,6 +1,6 @@
 ﻿using BlueprintCore.Actions.Builder;
-using BlueprintCore.Blueprints;
 using BlueprintCore.Conditions.Builder;
+using BlueprintCore.Utils;
 using Kingmaker.Blueprints;
 using Kingmaker.ElementsSystem;
 using System;